pyproject.toml
[project]
name = "orbit-ahp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ahp>=0.0.1",
    "google-generativeai>=0.8.5",
    "lib>=4.0.0",
    "matplot>=0.1.9",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "supabase>=2.23.0",
]
